chore(todo): resolve and remove TODO entries 250-257
- db2.cpp/h: clarify lazy-parse and sync-blocking adaptations (250, 251, 253)
- WDCReader.cpp: align getRelationRows !isLoaded error message with JS (252)
- dbd-manifest.cpp/h: drop jsonTruthy(), use uint32_t for FileDataIDs,
  remove stale TODO 191/192 comments (254-257)
- tab_data.cpp / casc-source.cpp: propagate uint32_t, drop redundant casts
- TODO_TRACKER.md: remove resolved entries 250-257 (367 remaining)

+++ src/js/casc/casc-source.cpp
@@ -224,7 +224,7 @@ BLTEReader CASC::getFileByName(const std::string& fileName, bool partialDecrypt,
 			std::string table_name = fileName.substr(14, fileName.length() - 14 - 4);
 			auto dbdId = dbd_manifest::getByTableName(table_name);
 			if (dbdId.has_value())
-				fileDataID = static_cast<uint32_t>(dbdId.value());
+				fileDataID = dbdId.value();
 		}
 
 		// fallback to listfile
@@ -298,7 +298,7 @@ BufferWrapper CASC::getVirtualFileByName(const std::string& fileName, bool suppr
 			std::string table_name = fileName.substr(14, fileName.length() - 14 - 4);
 			auto dbdId = dbd_manifest::getByTableName(table_name);
 			if (dbdId.has_value())
-				fileDataID = static_cast<uint32_t>(dbdId.value());
+				fileDataID = dbdId.value();
 		}
 
 		// fallback to listfile
